+++ galacteek/ld/rdf/terms/DID.py
@@ -1,3 +1,7 @@
 from rdflib import URIRef
 
+# IPFS CID of the latest document for an IPID
 documentIpfsCid = URIRef('https://w3id.org/did#documentIpfsCid')
+
+# Datetime to record when this DID was last "seen"
+didDateLastSeen = URIRef('https://w3id.org/did#dateLastSeen')